services/sentiment.py
# services/sentiment.py
from typing import Dict, List, Optional
from uuid import UUID
import openai
from models.memory import Emotion

class EmpatheticInterviewer:
    def __init__(self):
        self.openai = openai

    async def analyze_sentiment(self, text: str, language: str) -> Dict:
        try:
            response = await self.openai.ChatCompletion.create(
                model="gpt-4",
                messages=[
                    {"role": "system", "content": "Analyze the emotional content and sentiment of this text."},
                    {"role": "user", "content": text}
                ],
                functions=[{
                    "name": "analyze_emotion",
                    "parameters": {
                        "type": "object",
                        "properties": {
                            "emotions": {
                                "type": "array",
                                "items": {
                                    "type": "object",
                                    "properties": {
                                        "type": {"type": "string"},
                                        "intensity": {"type": "number"},
                                        "description": {"type": "string"}
                                    }
                                }
                            },
                            "requires_support": {"type": "boolean"},
                            "overall_sentiment": {"type": "string"}
                        }
                    }
                }]
            )
            return response.choices[0].message.function_call.arguments
        except Exception as e:
            raise ValueError(f"Sentiment analysis failed: {str(e)}")

    async def process_interview_response(
        self,
        profile_id: UUID,
        session_id: UUID,
        response: str,
        language: str
    ) -> Dict:
        sentiment = await self.analyze_sentiment(response, language)
        return {
            'sentiment': sentiment,
            'follow_up': await self._generate_empathetic_follow_up(sentiment, language)
        }

    async def _generate_empathetic_follow_up(self, sentiment: Dict, language: str):
        pass